airflow/requirements.txt
apache-airflow==2.8.0
psycopg2-binary>=2.9.9
polars-lts-cpu>=1.33.0
pyarrow>=12.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
